+++ CHANGELOG.md
@@ -1,3 +1,55 @@
-2013-11-29 - Changelog created
 
-2013-11-26 - v0.9.13-alpha: nav router back button fixes (remove back button on first route, fix forward/back transition on iOS devices which lack history.state)
+## 0.9.19 "Alpha Koala" (2014-01-09)
+ - Created ViewState Service to track navigation history
+ - Created navView directive
+ - Removed navPage and navRouter directives
+ - Using AngularUI Router instead of $route
+ - Update examples to use $stateProvider instead $routeProvider
+
+
+## 0.9.18 "Alpha Jaguar" (2013-12-18)
+ - Slide box bug fixes
+ - Fixed issues with minification
+ - Small tweaks
+
+
+## 0.9.17 "Alpha Iguana" (2013-12-16)
+ - Nav bar button fixes
+ - New slide box overhaul
+ - New list overhaul
+ - Radio button and checkbox fixes
+
+
+## 0.9.16 "Alpha Hippo" (2013-12-12)
+ - Scrollbars
+ - Scroll bug fixes
+
+
+## 0.9.15 "Alpha Giraffe" (2013-12-08)
+ - Scroll view fixes
+ - Radio button and checkbox fixes
+ - Slide box fixes
+
+
+## 0.9.14 "Alpha Fox" (2013-12-04)
+ - Massive scroll performance improvements
+ - Android fixes and perf improvements
+ - Header and nav bar button fixes
+ - Animation performance improvements
+
+
+## 0.9.13 "Alpha Elephant" (2013-11-26)
+ - nav router back button fixes
+ - Remove back button on first route
+ - Fix forward/back transition on iOS devices which lack history.state
+
+
+## 0.9.12 "Alpha Dog" (2013-11-24)
+ - Fixed some overflow scrolling issues.
+
+
+## 0.9.11 "Alpha Cat Dog" (2013-11-23)
+ 
+
+## 0.9.10 "Alpha Cat" (2013-11-23)
+